+++ server/src/main/java/eu/openanalytics/phaedra/plateservice/service/ProjectService.java
@@ -1,15 +1,20 @@
 package eu.openanalytics.phaedra.plateservice.service;
 
-import eu.openanalytics.phaedra.plateservice.model.Project;
-import eu.openanalytics.phaedra.plateservice.repository.ProjectRepository;
-import eu.openanalytics.phaedra.platservice.dto.ProjectDTO;
-import org.modelmapper.Conditions;
-import org.modelmapper.ModelMapper;
-import org.springframework.stereotype.Service;
-
+import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
+
+import org.modelmapper.Conditions;
+import org.modelmapper.ModelMapper;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import eu.openanalytics.phaedra.plateservice.model.Project;
+import eu.openanalytics.phaedra.plateservice.repository.ProjectRepository;
+import eu.openanalytics.phaedra.platservice.dto.ProjectDTO;
+import eu.openanalytics.phaedra.platservice.enumartion.ProjectAccessLevel;
+import eu.openanalytics.phaedra.util.auth.IAuthorizationService;
 
 @Service
 public class ProjectService {
@@ -19,30 +24,44 @@ public class ProjectService {
 	
 	private final ExperimentService experimentService;
 
-	public ProjectService(ProjectRepository projectRepository, ExperimentService experimentService) {
+	private final ProjectAccessService projectAccessService;
+	
+	@Autowired
+	private IAuthorizationService authService;
+	
+	public ProjectService(ProjectRepository projectRepository, ExperimentService experimentService, ProjectAccessService projectAccessService) {
 		this.projectRepository = projectRepository;
 		this.experimentService = experimentService;
+		this.projectAccessService = projectAccessService;
 	}
 
 	public ProjectDTO createProject(ProjectDTO projectDTO) {
+		projectAccessService.checkCanCreateProjects();
+		
 		Project project = new Project();
-		modelMapper.typeMap(ProjectDTO.class, Project.class)
-				.map(projectDTO, project);
+		modelMapper.typeMap(ProjectDTO.class, Project.class).map(projectDTO, project);
+		project.setCreatedBy(authService.getCurrentPrincipalName());
+		project.setCreatedOn(new Date());
 		project = projectRepository.save(project);
+		
 		return mapToProjectDTO(project);
 	}
 
 	public void updateProject(ProjectDTO projectDTO) {
 		Optional<Project> project = projectRepository.findById(projectDTO.getId());
 		project.ifPresent(p -> {
+			projectAccessService.checkAccessLevel(p.getId(), ProjectAccessLevel.Write);
 			modelMapper.typeMap(ProjectDTO.class, Project.class)
 					.setPropertyCondition(Conditions.isNotNull())
 					.map(projectDTO, p);
+			p.setUpdatedBy(authService.getCurrentPrincipalName());
+			p.setUpdatedOn(new Date());
 			projectRepository.save(p);
 		});
 	}
 
 	public void deleteProject(long projectId) {
+		projectAccessService.checkAccessLevel(projectId, ProjectAccessLevel.Admin);
 //		experimentService.deleteExperimentsByProjectId(projectId);
 		projectRepository.deleteById(projectId);
 	}
@@ -50,13 +69,17 @@ public class ProjectService {
 	public List<ProjectDTO> getAllProjects() {
 		List<Project> projects = (List<Project>) projectRepository.findAll();
 		return projects.stream()
+				.filter(p -> projectAccessService.hasAccessLevel(p.getId(), ProjectAccessLevel.Read))
 				.map(this::mapToProjectDTO)
 				.collect(Collectors.toList());
 	}
 	
 	public ProjectDTO getProjectById(long projectId) {
 		Optional<Project> result = projectRepository.findById(projectId);
-		return result.map(this::mapToProjectDTO).orElse(null);
+		return result
+				.map(this::mapToProjectDTO)
+				.filter(p -> projectAccessService.hasAccessLevel(p.getId(), ProjectAccessLevel.Read))
+				.orElse(null);
 	}
 
 	private ProjectDTO mapToProjectDTO(Project project) {
